Remove deprecated widget.Radio (use widget.RadioGroup instead)
