+++ widget/radio_group.go
@@ -209,3 +209,17 @@ func (r *radioGroupRenderer) updateItems() {
 		item.Refresh()
 	}
 }
+
+func removeDuplicates(options []string) []string {
+	var result []string
+	found := make(map[string]bool)
+
+	for _, option := range options {
+		if _, ok := found[option]; !ok {
+			found[option] = true
+			result = append(result, option)
+		}
+	}
+
+	return result
+}
